backend/requirements.txt
pandas
scikit-learn
nltk
fastapi
uvicorn
joblib
python-multipart
pymongo
passlib[bcrypt]
python-jose
